Combine the 3 ways to upload data into one, remove the dialog buttons
